feat: null_if_empty
